// prisma/generated/zod/inputTypeSchemas/CancellationRequestUpdateInputSchema.ts
import { Prisma } from '@prisma/client';

import { z } from 'zod';
import { StringFieldUpdateOperationsInputSchema } from './StringFieldUpdateOperationsInputSchema';
import { NullableStringFieldUpdateOperationsInputSchema } from './NullableStringFieldUpdateOperationsInputSchema';
import { CancellationSourceSchema } from './CancellationSourceSchema';
import { EnumCancellationSourceFieldUpdateOperationsInputSchema } from './EnumCancellationSourceFieldUpdateOperationsInputSchema';
import { CancellationStatusSchema } from './CancellationStatusSchema';
import { EnumCancellationStatusFieldUpdateOperationsInputSchema } from './EnumCancellationStatusFieldUpdateOperationsInputSchema';
import { NullableDateTimeFieldUpdateOperationsInputSchema } from './NullableDateTimeFieldUpdateOperationsInputSchema';
import { isValidDecimalInput } from './isValidDecimalInput';
import { DecimalJsLikeSchema } from './DecimalJsLikeSchema';
import { NullableDecimalFieldUpdateOperationsInputSchema } from './NullableDecimalFieldUpdateOperationsInputSchema';
import { DateTimeFieldUpdateOperationsInputSchema } from './DateTimeFieldUpdateOperationsInputSchema';
import { BookingUpdateOneRequiredWithoutCancellationRequestNestedInputSchema } from './BookingUpdateOneRequiredWithoutCancellationRequestNestedInputSchema';
import { BookingItemUpdateOneWithoutCancellationRequestNestedInputSchema } from './BookingItemUpdateOneWithoutCancellationRequestNestedInputSchema';

export const CancellationRequestUpdateInputSchema: z.ZodType<Prisma.CancellationRequestUpdateInput> = z.object({
  id: z.union([ z.string().cuid(),z.lazy(() => StringFieldUpdateOperationsInputSchema) ]).optional(),
  requestedBy: z.union([ z.string(),z.lazy(() => NullableStringFieldUpdateOperationsInputSchema) ]).optional().nullable(),
  reason: z.union([ z.string(),z.lazy(() => NullableStringFieldUpdateOperationsInputSchema) ]).optional().nullable(),
  source: z.union([ z.lazy(() => CancellationSourceSchema),z.lazy(() => EnumCancellationSourceFieldUpdateOperationsInputSchema) ]).optional(),
  status: z.union([ z.lazy(() => CancellationStatusSchema),z.lazy(() => EnumCancellationStatusFieldUpdateOperationsInputSchema) ]).optional(),
  processedAt: z.union([ z.coerce.date(),z.lazy(() => NullableDateTimeFieldUpdateOperationsInputSchema) ]).optional().nullable(),
  approvedBy: z.union([ z.string(),z.lazy(() => NullableStringFieldUpdateOperationsInputSchema) ]).optional().nullable(),
  approvedAt: z.union([ z.coerce.date(),z.lazy(() => NullableDateTimeFieldUpdateOperationsInputSchema) ]).optional().nullable(),
  declinedAt: z.union([ z.coerce.date(),z.lazy(() => NullableDateTimeFieldUpdateOperationsInputSchema) ]).optional().nullable(),
  refundAmount: z.union([ z.union([z.number(),z.string(),z.instanceof(Prisma.Decimal),DecimalJsLikeSchema,]).refine((v) => isValidDecimalInput(v), { message: 'Must be a Decimal' }),z.lazy(() => NullableDecimalFieldUpdateOperationsInputSchema) ]).optional().nullable(),
  notes: z.union([ z.string(),z.lazy(() => NullableStringFieldUpdateOperationsInputSchema) ]).optional().nullable(),
  createdAt: z.union([ z.coerce.date(),z.lazy(() => DateTimeFieldUpdateOperationsInputSchema) ]).optional(),
  updatedAt: z.union([ z.coerce.date(),z.lazy(() => DateTimeFieldUpdateOperationsInputSchema) ]).optional(),
  booking: z.lazy(() => BookingUpdateOneRequiredWithoutCancellationRequestNestedInputSchema).optional(),
  bookingItem: z.lazy(() => BookingItemUpdateOneWithoutCancellationRequestNestedInputSchema).optional()
}).strict();

export default CancellationRequestUpdateInputSchema;